requirements.txt
streamlit
pdfplumber
Pillow
requests
anthropic
openai
python-dotenv
pyperclip